Introduce ExecutableTransaction trait
