+++ transaction/src/builder/mod.rs
@@ -1,3 +1,5 @@
+mod manifest_builder;
 mod transaction_builder;
 
+pub use manifest_builder::ManifestBuilder;
 pub use transaction_builder::TransactionBuilder;
