+++ transaction/src/model/mod.rs
@@ -1,9 +1,5 @@
-mod instruction;
+mod executable;
 mod transaction;
-mod validated_instruction;
-mod validated_transaction;
 
 pub use self::transaction::*;
-pub use instruction::*;
-pub use validated_instruction::*;
-pub use validated_transaction::*;
+pub use executable::*;
