+++ transaction/src/validation/mod.rs
@@ -1,5 +1,9 @@
 mod id_allocator;
 mod id_validator;
+mod test_transaction;
+mod validated_transaction;
 
 pub use id_allocator::*;
 pub use id_validator::*;
+pub use test_transaction::*;
+pub use validated_transaction::*;
